função inserida da SymTable pelo SINTÁTICO. TODO: fazer isso para todo ID [que não estão mais sendo salvos!]

+++ trab/exp/test/simple/defFun.c
@@ -1,2 +1,2 @@
-int main(int p, float f){}
+float main(int p, float f){}
 int n(){}
